use-latest-values - move yes-no-unknown to yes-no.ts
