+++ src/app/user/(logged-in)/my-pets/view-dog/[dogId]/page.tsx
@@ -2,12 +2,8 @@ import { BarkButton } from "@/components/bark/bark-button";
 import { BarkDogAvatar } from "@/components/bark/bark-dog-avatar";
 import { BarkStatusBlock } from "@/components/bark/bark-status-block";
 import { getAuthenticatedUserActor } from "@/lib/auth";
-import {
-  DOG_ANTIGEN_PRESENCE,
-  DogAntigenPresence,
-  YES_NO_UNKNOWN,
-  YesNoUnknown,
-} from "@/lib/data/db-enums";
+import { DOG_ANTIGEN_PRESENCE, DogAntigenPresence } from "@/lib/data/db-enums";
+import { YES_NO_UNKNOWN, YesNoUnknown } from "@/lib/bark/enums/yes-no";
 import { DOG_GENDER, DogGender } from "@/lib/bark/models/dog-gender";
 import {
   DogAppointment,
