+++ src/app/user/registration/_components/donor-form.tsx
@@ -16,7 +16,8 @@ import { useRouter } from "next/navigation";
 import { BarkFormOption } from "@/components/bark/bark-form";
 import { postRegistrationRequest } from "@/app/user/registration/_actions/post-registration-request";
 import { RegistrationRequest } from "@/lib/services/registration";
-import { DogAntigenPresence, YesNoUnknown } from "@/lib/data/db-enums";
+import { DogAntigenPresence } from "@/lib/data/db-enums";
+import { YesNoUnknown } from "@/lib/bark/enums/yes-no";
 import { DogGender } from "@/lib/bark/models/dog-gender";
 import { UserResidency } from "@/lib/data/db-enums";
 import { BARK_UTC } from "@/lib/utilities/bark-time";
