+++ src/lib/admin/admin-models.ts
@@ -1,4 +1,4 @@
-import { YesNoUnknown } from "../data/db-enums";
+import { YesNoUnknown } from "../bark/enums/yes-no";
 import { DogGender } from "../bark/models/dog-gender";
 
 export type IncompleteProfile = {
