+++ src/lib/services/registration.ts
@@ -9,7 +9,8 @@ import {
   DogSecureOii,
   DogSpec,
 } from "@/lib/data/db-models";
-import { YesNoUnknown, DogAntigenPresence } from "../data/db-enums";
+import { DogAntigenPresence } from "../data/db-enums";
+import { YesNoUnknown } from "../bark/enums/yes-no";
 import { DogGender } from "../bark/models/dog-gender";
 import { UserResidency } from "../data/db-enums";
 import { dbSelectUserIdByHashedEmail, dbInsertUser } from "@/lib/data/db-users";
