+++ src/lib/user/actions/get-dog-profile.ts
@@ -2,7 +2,8 @@ import { dbResultQuery } from "@/lib/data/db-utils";
 import { UserActor } from "../user-actor";
 import { DogProfile } from "@/lib/dog/dog-models";
 import { Err, Ok, Result } from "@/lib/utilities/result";
-import { DogAntigenPresence, YesNoUnknown } from "@/lib/data/db-enums";
+import { DogAntigenPresence } from "@/lib/data/db-enums";
+import { YesNoUnknown } from "@/lib/bark/enums/yes-no";
 import { DogGender } from "@/lib/bark/models/dog-gender";
 import { CODE } from "@/lib/utilities/bark-code";
