+++ src/lib/vet/vet-models.ts
@@ -1,4 +1,4 @@
-import { YesNoUnknown } from "../data/db-enums";
+import { YesNoUnknown } from "../bark/enums/yes-no";
 import { DogGender } from "../bark/models/dog-gender";
 
 // WIP: Define CallTask and EncryptedCallTask in lib/bark/models based fields in AvailableDog and OwnerContactDetails.
