+++ tests/_fixtures.ts
@@ -20,12 +20,8 @@ import {
   DbCallSpec,
   DbCallGen,
 } from "@/lib/data/db-models";
-import {
-  DOG_ANTIGEN_PRESENCE,
-  DogAntigenPresence,
-  YES_NO_UNKNOWN,
-  YesNoUnknown,
-} from "@/lib/data/db-enums";
+import { DOG_ANTIGEN_PRESENCE, DogAntigenPresence } from "@/lib/data/db-enums";
+import { YES_NO_UNKNOWN, YesNoUnknown } from "@/lib/bark/enums/yes-no";
 import { DOG_GENDER, DogGender } from "@/lib/bark/models/dog-gender";
 import { USER_RESIDENCY } from "@/lib/data/db-enums";
 import { dbInsertAdmin, dbSelectAdmin } from "@/lib/data/db-admins";
