+++ tests/admin/get-num-incomplete-profiles.test.ts
@@ -7,7 +7,8 @@ import {
   insertUser,
 } from "../_fixtures";
 import { Pool } from "pg";
-import { PARTICIPATION_STATUS, YES_NO_UNKNOWN } from "@/lib/data/db-enums";
+import { PARTICIPATION_STATUS } from "@/lib/data/db-enums";
+import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no";
 import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
 import {
   DEFAULT_DATE_TIME_FORMAT,
