+++ tests/user/add-my-dog.test.ts
@@ -7,7 +7,8 @@ import {
   insertVet,
 } from "../_fixtures";
 import { UTC_DATE_OPTION, parseDateTime } from "@/lib/utilities/bark-time";
-import { DOG_ANTIGEN_PRESENCE, YES_NO_UNKNOWN } from "@/lib/data/db-enums";
+import { DOG_ANTIGEN_PRESENCE } from "@/lib/data/db-enums";
+import { YES_NO_UNKNOWN } from "@/lib/bark/enums/yes-no";
 import { DOG_GENDER } from "@/lib/bark/models/dog-gender";
 import { addMyDog } from "@/lib/user/actions/add-my-dog";
